pyproject.toml
[tool.poetry]
name = "shadowmessenger"
version = "0.1.3"
description = "256bit AES CFB end to end encrypted messenger over ip"
authors = ["ShadowDrums <122622018+ShadowDrums@users.noreply.github.com>, DJ Stomp <85457381+DJStompZone@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pycryptodome = ">=3.20.0,<3.21.0"


[tool.poetry.group.dev.dependencies]
poetry = "^1.8.3"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
